pyproject.toml
[project]
name = "firsfastapi"
version = "0.1.0"
description = ""
authors = [
    {name = "Shoxrux Yuldashov",email = "deagleshox@mail.ru"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi (>=0.115.7,<0.116.0)",
    "uvicorn (>=0.34.0,<0.35.0)",
    "httpie (>=3.2.4,<4.0.0)",
    "requests (>=2.32.3,<3.0.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "mypy (>=1.14.1,<2.0.0)",
    "faker (>=35.0.0,<36.0.0)",
    "pytest (>=8.3.4,<9.0.0)",
    "opentelemetry-api (>=1.29.0,<2.0.0)",
    "opentelemetry-sdk (>=1.29.0,<2.0.0)",
    "opentelemetry-distro (>=0.50b0,<0.51)",
    "opentelemetry-exporter-otlp (>=1.29.0,<2.0.0)",
    "prometheus-api-client (>=0.5.5,<0.6.0)",
    "pyroscope-io (>=0.8.8,<0.9.0)",
    "python-decouple (>=3.8,<4.0)",
    "python-jose (>=3.3.0,<4.0.0)",
    "passlib (>=1.7.4,<2.0.0)",
    "python-multipart (>=0.0.20,<0.0.21)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
